[Refactor] Rename HomeActivity To MainHostActivity and add AppNavigator to represent new navigation arch
