+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".presentation.activity.HomeActivity"
+        <activity android:name=".presentation.activity.MainHostActivity"
             android:windowSoftInputMode="adjustResize"
             android:exported="true">
             <intent-filter>
